Change Double with Bigdecimal for calculation, add test and rename the methods and names

+++ src/main/scala/bank/model/Operation.scala
@@ -2,7 +2,7 @@ package bank.model
 
 import java.util.Date
 
-sealed abstract class Operation(value: PositifNumber, balanceBeforeOp: Double, date: Date) {
+sealed abstract class Operation(value: PositifNumber, balanceBeforeOp: BigDecimal, date: Date) {
   override def toString: String = s"value => $value ; date => $date; balance before operation => $balanceBeforeOp"
 }
 
@@ -10,7 +10,7 @@ sealed abstract class Operation(value: PositifNumber, balanceBeforeOp: Double, d
 object Operation {
 
   case class Withdrawal(value: PositifNumber,
-                        balanceBeforeOp: Double,
+                        balanceBeforeOp: BigDecimal,
                         date: Date) extends Operation(value, balanceBeforeOp, date) {
 
     override def toString: String = s"Withdrawal( ${super.toString} )"
@@ -18,7 +18,7 @@ object Operation {
   }
 
   case class Deposit(value: PositifNumber,
-                     balanceBeforeOp: Double,
+                     balanceBeforeOp: BigDecimal,
                      date: Date) extends Operation(value, balanceBeforeOp, date) {
 
     override def toString: String = s"Deposit( ${super.toString} )"
